部分服务器端代码，并行性最好放在网络层，update中回调回来

+++ NetworkDemo/ClientDemo/MessageClient.cs
@@ -36,7 +36,7 @@ public class MessageClient
         m_netlayerClient.m_disconnectEvent += this.OnDisconnct;
     }
 
-    public void OnUpdate(float delta)
+    public void OnUpdate(int delta)
     {
         m_netlayerClient.OnUpdate(delta);
